+++ js/interactions/index.js
@@ -7,3 +7,4 @@ export { DragHandler } from './DragHandler.js';
 export { ResizeHandler } from './ResizeHandler.js';
 export { RotateHandler } from './RotateHandler.js';
 export { AlignmentGuides } from './AlignmentGuides.js';
+export { MarqueeHandler } from './MarqueeHandler.js';